Allows for "expand" option in search requests.
Since this utilizes NULL for the default, it should not create any BC breaks. Also, the underlying HTTP transmission code utilizes "http_build_query()" from PHP, and in this function array entries with NULL values are not included in the output. So, this change wouldn't even affect the URLs that are generated.