Added phpcs to composer and used more granular @dev-directive instead of global minimum-stability setting
Since we require devs to run the coding standard we might as well make it as easy as possible by adding it to the require-dev dependencies, just like we do with phpunit. Updated the CONTRIBUTING.md to reflect this.
Also I think it's bad practise to se the minimum-stability setting of all your dependencies to dev, since then we also get the dev builds of dependencies that have tested, stable releases. This change ensures the latest master of the CodingStyle, but the latest stable release of other libs.